Melhorando tabela de simbolos, e encapsulando valores (#22,#25). Necessita mudar os enums da arvore mais uma vez(#22)
