+++ src/datastructures/STable.c
@@ -6,14 +6,15 @@
 
 
 // entry
-TableEntry* createEntry(char* name, int type, char* val, int methodParam, STable* parentScope) {
+TableEntry* createEntry(char* name, Object* val, int methodParam, STable* parentScope) {
     
     TableEntry* newEntry = (TableEntry*) malloc(sizeof(TableEntry));
-    newEntry->name = name;
-    newEntry->type = type;
-    newEntry->val = val;
+    newEntry->name = name; // note que para strings e para valores void criados (talvez seja necessário malloc)
+    newEntry->val = val;   // objects sempre vão ser alocados, string devemos ter tratativa! (a não ser que essa seja alocada pelo bison?)
     newEntry->methodParam = methodParam;
     newEntry->parentScope = parentScope;
+    newEntry->level = parentScope->level;
+  	newEntry->order = parentScope->order;
     
     
     return newEntry;
@@ -21,29 +22,27 @@ TableEntry* createEntry(char* name, int type, char* val, int methodParam, STable
     
 }
 
-void printEntry(TableEntry* e) {
-    printf("%s : (%s, %s, methodParam: %d ) \n",e->name,mappingEnumEntry[e->type],e->val,e->methodParam);
-    if(e->parentScope){
-        printTable(e->parentScope);
-    }
-    
-}
 
-void letgoEntry(TableEntry* e){
-	if(!e) {
-	    return;
+void printEntry(TableEntry* e) {
+
+	if(e) 
+	{
+		printf("\t ");
+		//printf("\t %s : ( ",e->name);
+		prinObject(e->val);
+		printf("\t ( methodParam: %d, level: %d, order: %d ) ",e->methodParam,e->level,e->order);
+
 	}
-	free(e);
+
 }
 
 
 
 /// table
 
-STable* createTable(int type, STable* parent,  int level, int order) {
+STable* createTable(SCOPE_TYPE type, STable* parent,  int level, int order) {
 
 	STable* newtable = (STable*) malloc(sizeof(STable));
-	
 	
 /*	
 	int nt = sizechildren /sizeof (STable*);
@@ -58,6 +57,11 @@ STable* createTable(int type, STable* parent,  int level, int order) {
 	newtable->type = type;
 	newtable->level = level;
 	newtable->order = order;
+
+	newtable->type = type;
+	newtable->level = level;
+	newtable->order = order;
+	newtable->lastEntryIndex = 0;
 	
 	if(parent){
 		newtable->parent = parent;
@@ -73,44 +77,27 @@ STable* createTable(int type, STable* parent,  int level, int order) {
 
 	return newtable;
 
-
-
 }
 
 void printTable(STable* t){
 	if(t){
 		printf("%s (%d,%d) \n",mappingEnumTable[t->type],t->level,t->order);
+		if(t->tableData){
+			printf("|--> Entries: \n");
+			int i;
+			for(i=0;i<= t->lastEntryIndex; i++){
+				printEntry(t->tableData[i]);
+			}
+		}
 	}
 }
 
 
-
-int hash(const char * str) {
-	int hash = 401;
-	int c;
-
-	while (*str != '\0') {
-		hash = ((hash << 4) + (int)(*str)) % MAX_TABLE;
-		str++;
+void letgoEntry(TableEntry* e){
+	if(!e) {
+	    return;
 	}
-
-	return hash % MAX_TABLE;
-}
-
-void insert(STable* t, TableEntry* e) {
-    
-    int index = hash(e->name);
-    printf("HASH CALCULADO É: %d \n",index);
-    t->tableData[index] = e;
-    
-    
-}
-
-TableEntry* lookup(STable* t, const char* name) {
-    
-    int index = hash(name);
-    return t->tableData[index];
-    
+	free(e);
 }
 
 
@@ -135,3 +122,73 @@ void letgoTable(STable* t){
 	}
 	free(t);
 }
+
+
+
+
+
+
+int hash(const char * str) {
+	int hash = 401;
+	int c;
+
+	while (*str != '\0') {
+		hash = ((hash << 4) + (int)(*str)) % MAX_TABLE;
+		str++;
+	}
+
+	return hash % MAX_TABLE;
+}
+
+void insert(STable* t, TableEntry* e) {
+    
+    int index = hash(e->name);
+    printf("HASH CALCULADO para (%s) É: %d \n",e->name,index);
+   	t->tableData[index] = e;
+    t->lastEntryIndex = index > t->lastEntryIndex ? index : t->lastEntryIndex;
+    
+}
+
+TableEntry* lookup(STable* t, const char* name) {
+    
+    int index = hash(name);
+    return t->tableData[index];
+    
+}
+
+
+
+// op's novas que funcionam
+
+// note que ! eu poderia muito bem por que os valores FOLHA dos nós da arvore são os mesmos que ENTRY_TYPE, existe ligeira redundância...
+// QUE POR SINAL.... meu entry_type vai ser o meu "proxy" para valores? 
+		// acho que vai ter que ter proxy... senão vai ficar dificil manipular os TIPOS
+		// JÁ QUE OS TIPOS VÃO SER LITERALMENTE REALCIONADOS A DESEMPILHAR A ARVORE
+		// primeiro eu vou testar usando valores literais 
+void addValueCurrentScope(char* name, ENTRY_TYPE type, void* val, int methodParam,STable* current) {
+
+	//nome da variavel,   val vai ser literalmente o valor dela (problema, e quanto for uma lista?)
+	TableEntry* entry = createEntry(name,type,val,methodParam,current);
+	insert(current,entry);
+}
+
+
+
+
+
+
+//void ou retorna o filho??? Retornar parece melhor, provavelmente ao adicionar um subscope eu vou querer operar sobre ele imediatamente
+STable* addSubScope(STable* parent, SCOPE_TYPE type) {
+	
+	STable* child = createTable(type,parent,parent->level+1,parent->nchild);
+
+	if(!parent->nchild) {
+		parent->children = (STable**) malloc(parent->nchild*sizeof(STable*));
+	}
+	parent->children[parent->nchild] = child;
+	parent->nchild++;
+
+	return child;
+
+}
+
